refactor: separate /*info commands from ToolsCommandGroup
